Exempt Preview Builds from Prerelease Filter
We currently blanket ignore releases with a prerelease segment. This change makes it so that we add branches preview releases (tagged with "preview" in their prerelease segment). This adds some noise, but does not add canary builds.

I did a quick look at `new-release.js` to see if we might have logic that is not friendly to versions with prerelease segment, but it looks like it should work on first glace.